Do not fail on physical interface update
When creating or updating interface, if the interface name does not
match the interface type, operation fails. However, the check logic is
not suited for update operation, since it does not recognize physical
interfaces (we support just updates of physical interfaces). This way
updating of physical interface always fails. Moreover, it is not really
needed to perform this check upon update, because we cannot change
interface type, once the interface is already created.

Do not validate interface name against interface type during update
operation.

Change-Id: If8ddd872ddca4682772a00e8e2ea24543e19b905
Signed-off-by: Jakub Morvay <jmorvay@frinx.io>