UNI-601 : fixed using correct interface ID
YANG model stated 'id' field to be the key in interface list.
This also means the 'id' field MUST be filled in the config container.

        leaf id {
          type leafref {
            path "../config/id";
          }
          description
            "A pointer to the identifier for the interface.";
        }

Writer was written in a way that it used only the 'id' field from
this leaf (which resulted in a more complicated way to get the information),
therefore the tests didn't fail.

Signed-off-by: Dana Kutenicsova <dkutenicsova@frinx.io>
Change-Id: I571b5874e822557b1240cd0d04af319143cca8cc