FRHD-452 : fixed empty parsing
Some readers do not require all the fields to
be parsed. In case of base units, the pattern
used when no matching was needed, was "" which
resulted in false positive when the output contained
eg. newline (junos).

Fix this by expecting the subclasses override of the
parsing methods only when they actually provide a pattern.

This only concerns classes inherited from base units.

Signed-off-by: Dana Kutenicsova <dkutenicsova@frinx.io>
Change-Id: Ia894f24939e59918f23c1de6d9cf590d3efab983