Fixed is-in-sync check for ios and ios xe devices
Fixed problem with regex for getting data about time from device changes history.
We might want to look into the fact, that when the device history is clean (not sure how often that resets), the is-in-sync request might end in an error, stating that parsing of configuration fingerprint is not implemented for this device type.
This is a little misleading.

Signed-off-by: mmartiska <mmartiska@frinx.io>
Change-Id: Ie32cbaffa85ad45e77af7f45df3d2c9b80a83b49