[FRINX-1404] IOSXE deleting all serv instances
By deleting all service instances we might delete also the ones
that must not be deleted and shut down whole device.
I created a solution, where only those service instances
that are missing in dataAfter will be deleted,
and those that are going to be modified will just be rewritten.
This writer was not implemented in the chunk template way
but during the implementation I needed so make some operations with the data,
that the chunk template was not capable of so I had to keep the
original implementation and  just modify it.

Signed-off-by: mmartiska <mmartiska@frinx.io>
Change-Id: I8f34ec38131169084fbd6c7cb67651d76bfc1247